+++ .github/workflows/check.yml
@@ -1,0 +1,20 @@
+name: check
+on: pull_request
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v1
+      - name: Use Node.js 18
+        uses: actions/setup-node@v1
+        with:
+          node-version: 18
+      - name: install
+        run: |
+          yarn install
+      - name: build
+        run: |
+          yarn build
+      - name: export
+        run: |
+          yarn export
