--- .prettierrc
@@ -1,12 +1,0 @@
-{
-  "printWidth": 80,
-  "trailingComma": "all",
-  "overrides": [
-    {
-      "files": "*.js",
-      "options": {
-        "trailingComma": "es5"
-      }
-    }
-  ]
-}
