--- .travis.yml
@@ -1,16 +1,0 @@
-language: node_js
-node_js:
-  - "node"
-cache:
-  yarn: true
-script:
-  - "yarn lint"
-  - "yarn test"
-  - "yarn build"
-deploy:
-  provider: pages
-  skip_cleanup: true
-  github_token: $GITHUB_TOKEN
-  local_dir: build/dist
-  on:
-    branch: master
