+++ .vscode/extensions.json
@@ -1,7 +1,3 @@
 {
-  "recommendations": [
-    "EditorConfig.EditorConfig",
-    "eg2.tslint",
-    "esbenp.prettier-vscode"
-  ]
+  "recommendations": ["esbenp.prettier-vscode"]
 }
