--- .vscode/settings.json
@@ -1,6 +1,0 @@
-{
-  "prettier.printWidth": 80,
-  "prettier.singleQuote": false,
-  "prettier.trailingComma": "all",
-  "prettier.typescriptEnable": ["typescript", "typescriptreact"]
-}
