+++ README.md
@@ -1,3 +1,1 @@
 # nyanpass
-
-[![Build Status](https://travis-ci.org/kdama/nyanpass.svg?branch=master)](https://travis-ci.org/kdama/nyanpass)
