--- jest.config.json
@@ -1,12 +1,0 @@
-{
-  "collectCoverageFrom": ["src/scripts/**/*.{ts,tsx}"],
-  "coverageDirectory": "build/test/coverage",
-  "coverageReporters": ["cobertura", "html", "text-summary"],
-  "moduleFileExtensions": ["ts", "tsx", "js"],
-  "setupFiles": ["<rootDir>/__tests__/polyfill.ts"],
-  "testPathIgnorePatterns": ["/build/", "/node_modules/"],
-  "testRegex": "/__tests__/.*\\.test\\.(ts|tsx)$",
-  "transform": {
-    ".(ts|tsx)": "ts-jest"
-  }
-}
