+++ next.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  assetPrefix: process.env.GITHUB_PAGES ? "/nyanpass" : "/",
+};
