+++ package.json
@@ -8,49 +8,23 @@
     "url": "https://github.com/kdama/nyanpass.git"
   },
   "scripts": {
-    "start": "npm-run-all -l clean -p serve watch",
-    "test": "run-s test:clean test:single",
-    "build": "webpack",
-    "clean": "del-cli ./build",
-    "lint": "run-p lint:*",
-    "fix": "prettier -l --print-width 80 --trailing-comma all --write \"*.{ts,tsx}\" \"src/**/*.{ts,tsx}\" \"__tests__/**/*.{ts,tsx}\"",
-    "serve": "webpack-dev-server --inline --port 8000",
-    "sloc": "sloc ./src",
-    "watch": "run-s test:clean test:run",
-    "lint:prettier": "prettier -l --print-width 80 --trailing-comma all \"*.{ts,tsx}\" \"src/**/*.{ts,tsx}\" \"__tests__/**/*.{ts,tsx}\"",
-    "lint:tslint": "tslint \"./**/*.ts?(x)\" -e \"./node_modules/**\" -t stylish",
-    "test:clean": "del-cli ./build/test",
-    "test:run": "jest -c ./jest.config.json --watch",
-    "test:single": "jest -c ./jest.config.json --coverage"
+    "dev": "next",
+    "build": "next build",
+    "start": "next start",
+    "export": "next export"
   },
   "dependencies": {},
   "devDependencies": {
-    "@types/html-webpack-plugin": "^3.2.0",
-    "@types/jest": "^23.3.5",
-    "@types/node": "^10.11.7",
-    "@types/react": "^16.4.16",
-    "@types/react-dom": "^16.0.9",
-    "@types/react-test-renderer": "^16.0.3",
-    "@types/webpack": "^4.4.16",
-    "del-cli": "^1.1.0",
-    "html-webpack-plugin": "^3.2.0",
-    "jest": "^23.6.0",
-    "npm-run-all": "^4.1.3",
-    "prettier": "^1.14.3",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "remap-istanbul": "^0.12.0",
-    "ts-jest": "^23.10.4",
-    "ts-loader": "^5.2.1",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.11.0",
-    "tslint-react": "^3.6.0",
-    "typescript": "^3.1.3",
-    "webpack": "^4.20.2",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.9"
+    "@types/jest": "^29.5.5",
+    "@types/node": "^16.10.2",
+    "@types/react": "^17.0.26",
+    "jest": "^29.7.0",
+    "next": "^12.3.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "typescript": "^4.4.3"
   },
   "engines": {
-    "node": ">=6"
+    "node": ">=14"
   }
 }
