--- src/scripts/constants/Presets.ts
@@ -1,5 +1,0 @@
-export enum Presets {
-  Numbers = "0123456789",
-  LowerAlphabets = "abcdefghijklmnopqrstuvwxyz",
-  UpperAlphabets = "ABCDEFGHIJKLMNOPQRSTUVWXYZ",
-}
