--- src/scripts/types/Options.ts
@@ -1,4 +1,0 @@
-export interface Options {
-  includes: string[];
-  maxLength: number;
-}
