+++ tsconfig.json
@@ -1,14 +1,24 @@
 {
   "compilerOptions": {
-    "module": "commonjs",
-    "target": "es3",
-    "noImplicitAny": true,
+    "jsx": "preserve",
+    "lib": ["dom", "esnext"],
+    "module": "esnext",
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "sourceMap": true,
-    "jsx": "react",
-    "strictNullChecks": true,
-    "lib": ["es2017", "dom"]
+    "strict": true,
+    "target": "es5",
+    "allowJs": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "noEmit": true,
+    "esModuleInterop": true,
+    "moduleResolution": "node",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "incremental": true
   },
-  "compileOnSave": false
+  "compileOnSave": false,
+  "exclude": ["node_modules"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"]
 }
